Add an option whether to check updates or not

+++ ogb/graphproppred/dataset.py
@@ -7,7 +7,7 @@ from ogb.io.read_graph_raw import read_csv_graph_raw, read_binary_graph_raw
 import torch
 
 class GraphPropPredDataset(object):
-    def __init__(self, name, root = 'dataset', meta_dict = None):
+    def __init__(self, name, root = 'dataset', meta_dict = None, check_update = True):
         '''
             - name (str): name of the dataset
             - root (str): root directory to store the dataset folder
@@ -41,7 +41,7 @@ class GraphPropPredDataset(object):
         # First check whether the dataset has been already downloaded or not.
         # If so, check whether the dataset version is the newest or not.
         # If the dataset is not the newest version, notify this to the user. 
-        if osp.isdir(self.root) and (not osp.exists(osp.join(self.root, 'RELEASE_v' + str(self.meta_info['version']) + '.txt'))):
+        if check_update and osp.isdir(self.root) and (not osp.exists(osp.join(self.root, 'RELEASE_v' + str(self.meta_info['version']) + '.txt'))):
             print(self.name + ' has been updated.')
             if input('Will you update the dataset now? (y/N)\n').lower() == 'y':
                 shutil.rmtree(self.root)
